# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
  description: 'TaskY API — canonical contract generated from springdoc on a dev-profile boot (Observed). Regenerate: boot
    api with dev profile, GET /api-docs, convert to YAML.'
  x-generated: 2026-09-26, feat OIDC code exchange + timer e2e
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/auth/switch-org:
    post:
      tags:
      - auth-controller
      operationId: switchOrg
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchOrgRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SwitchOrgResponse'
...
  /api/v1/auth/oidc:
    post:
      tags:
      - auth-controller
      operationId: loginWithOidc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OidcAuthRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/oidc/code:
    post:
      tags:
      - auth-controller
      operationId: loginWithOidcCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OidcCodeRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /api/v1/auth/google:
    post:
      tags:
      - auth-controller
      operationId: loginWithGoogle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  schemas:
...
    SwitchOrgRequest:
      type: object
      properties:
        orgId:
          type: string
          format: uuid
      required:
      - orgId
    OrgInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        role:
          type: string
        timezone:
          type: string
        workWeekStartsOn:
          type: integer
          format: int32
    SwitchOrgResponse:
      type: object
      properties:
        token:
          type: string
        org:
          $ref: '#/components/schemas/OrgInfo'
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
    OidcAuthRequest:
      type: object
      properties:
        provider:
          type: string
          enum:
          - GOOGLE
          - MICROSOFT
          - MOCK_GOOGLE
          - MOCK_MICROSOFT
        idToken:
          type: string
          minLength: 1
      required:
      - provider
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserInfo'
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/OrgInfo'
    OidcCodeRequest:
      type: object
      properties:
        provider:
          type: string
          enum:
          - GOOGLE
          - MICROSOFT
          - MOCK_GOOGLE
          - MOCK_MICROSOFT
        code:
          type: string
          minLength: 1
        codeVerifier:
          type: string
          minLength: 1
        redirectUri:
          type: string
          minLength: 1
      required:
      - provider
    AuthRequest:
      type: object
      properties:
        idToken:
          type: string
          minLength: 1
...